Harvest selectivity based on fish size and mesh size

Internal function to calculate the proportion of fish retained by a gillnet of a given mesh size, using the selectivity model described by Bromaghin (2005).

Usage

.calc_selectivity(size, meshsize, s)

Arguments

size

Numeric. Fish size in millimeters.

meshsize

Numeric. Stretch mesh size in inches.

s

Numeric. Standard deviation of selectivity, typically 0.204.

Value

Numeric. Proportion of fish of the given size retained by the net.

References

Bromaghin, J. F. (2005). The statistical evaluation of relative gillnet selectivity estimates and their application to the management of Alaskan sockeye salmon. Alaska Fishery Research Bulletin, 11(2), 258–266.

Examples

.calc_selectivity(size = 550, meshsize = 5.0, s = 0.204)
#> Error in .calc_selectivity(size = 550, meshsize = 5, s = 0.204): could not find function ".calc_selectivity"